refactor: add comments + move projects file in right place
