+++ webapi/src/test/scala/org/knora/webapi/responders/admin/ProjectsResponderADMSpec.scala
@@ -49,7 +49,6 @@ import org.knora.webapi.messages.admin.responder.usersmessages.{
   Keywords,
   Logo,
   Longname,
-  ProjectCreatePayloadADM,
   Selfjoin,
   Shortcode,
   Shortname,
